Remove uninteresting tests.
We have hundreds of queries in the test suite that use generate_series,
to generate test data. I don't see anything special in these queries to
justify their existence.

Likewise, the CREATE commands being removed are completely straightforward,
and we have a lot of similar commands in the test suite.