Speed up cs-uao TINC tests related to merge join
On gpdb4, the enable_mergejoin=off GUC was ignored, which allowed
certain tests to use merge join, running very quickly.  On gpdb5, this
GUC is now read correctly, which causes these tests to NOT use merge
join, which resulted in test runs taking up to 10 minutes longer. This
commit fixes the tests to properly set the GUC when attempting to test
merge join.

https://github.com/greenplum-db/gpdb/issues/1341