Remove only marginally useful mock test.
Digging into the git history, this was added as a test for the bugfix in
upstream commit 0899556e. However, when I tried reverting that commit, and
ran the test, it passed anyway. So it's not reliably testing the underlying
issue. In any case, it seems highly unlikely that the exact same bug would
be reintroduced, and the surrounding code has changed since, and the
upstream has been happy without an explicit case for this, let's just remove
the test rather than try to fix it.