Rewrite test for gp_max_plan_size.
This can easily be tested as a normal pg_regress-style test, no need
for a mock test. Mock tests are somewhat expensive, as they take a few
seconds to compile and link each. They are more lines of code and
more fragile to surrounding code changes, too.

The original test tried to do something clever with INT_MAX, but that
doesn't seem too interesting, given the current implementation. The
query plan is built in a StringData, which is limited to MaxAllocSize
(1 GB), so you'll get an "request too large" error from before you get
anywhere near INT_MAX.