Remove broken optfunctional regression test suite.
It was broken literally the next day that it was added to the repository,
by the 8.3 merge. Per the authors of the test suite, it is a known issue,
and there are no immediate plans to fix it. Remove it, to avoid confusion.

See discussion on gpdb-dev, thread '"Optimizer functional" tests':
https://groups.google.com/a/greenplum.org/d/msg/gpdb-dev/xRihGh6fAjQ/7-NyjJbFAAAJ