Guard against possible NULL pointer dereferencing
Improves defensiveness of programming around pointer derefencing to
ensure that we don't risk a NULL pointer. Most of these are quite
straight-forward, those of note are discussed below.

In doDispatchDtxProtocolCommand() we relied on the result data being
created in zeroed out memory on CdbDispatchDtxProtocolCommand() which
isn't guaranteed for every compiler. Explcitly set numResults to zero
and also check the results for NULL.

Per multiple reports by Coverity