Replace strncpy() uses with strlcpy()
Handle potential non-termination of buffer in OpenNamedFile. In case
the requested filename overflows the buffer size and gets truncated
(albeit unlikely), error out and log.

In fe-protocol3.c the strncpy() calls have been replaced with safer
strlcpy() calls. Sync the gp_libpq_fe sources to use the same.

Use strlcpy() in unit tests not becuase the uses were dangerous but
it doesn't hurt and in doing so we pass static analysis without having
to resort to lots of overrides. While a bit of code golf, passing
analysis clean is a good thing.

Per report from Coverity.