Guard against using uninitialized value
We only pass in the filename in the entry struct, which is used as
the key in the hash lookup. The mirrorBufpoolResyncChangedPageCount
member is inspected as well in the FileRepResync_UpdateEntry() code
however so set to zero as it will at worst lead to a no-op since the
the hashtable entry is only updated if zero (thus updating it from
zero to zero). The current coding relies on the struct being zeroed
out at definition which is shaky.

Per report by Coverity.