Add "installcheck-world" target (and sync plperl with upstream) (#1343)
* Sync plperl with PostgreSQL REL9_1_STABLE.

We had mostly backported plperl from 9.1, but not up to the tip of
the stable branch, and there were some cosmetic and other differences
that were not really necessary. It is simpler if the GPDB version is
as identical as possible to a particular upstream version, not a mismatch
of different versions.

Copy over everything from upstream REL9_1_STABLE, and tweak things to make
it compile and pass the regression tests.

I started hacking this because the regression tests were not passing. Now
they pass.

* Add "installcheck-world" makefile target, remove other top-level targets.

The master plan is that we will have straightforward "world" makefile
targets, like newer versions of PostgreSQL has, that build, install,
and test everything in the source tree. For example, "make world" builds
everything and the kitchen sink, and "make installcheck-world" runs
every test suite against an existing installation.

This patch is the first step in that direction. It creates a top-level
"installcheck-world" target, that runs some of the test suites we have.
It does not run everything, because the tests for some components are
either broken, or require additional setup to run. The point of this
first step is to establish the precedence, and clean up and add
everything else that we care about to this new target. For example, this
does not run the gphdfs tests, gpmapreduce tests, or orafce tests yet.

Also, this only adds the "installcheck-world" target, not "world",
"install-world" nor "check-world". Those are TODOs.

Remove other, defunct, top-level targets, to reduce confusion.

Move unittest-check top-level target from gpAux/Makefile to top-level
Makefile, to get some more visibility to it.

Adjust the concourse pipeline scripts to run the new "installcheck-world"
target, replacing the separate "installcheck-good" and
"installcheck-bugbuster" tasks. In the future, whenever a new test suite
is added, it can be added to the installcheck-world target, and the
pipeline will pick it up immediately, without having to modify the yaml
files.

An immediate benefit of this is that we now run the isolationtester test
suite, and the PL tests, as part of the pipeline.

* Fixed test cases