Don't report drop errors during postdata restore. (#1395)
When performing a dump, the postdata file contains DROP statements that will report an error if the relation already exists. This is misleading since the restore will say that errors have occured during restore, even though the DB is still in the expected state. For most cases we simply could use the `DROP IF EXISTS` syntax; however, the `DROP CONSTRAINT IF EXISTS` syntax isn't available until postgres 9.0 so we create a temporary procedure to perform this action.

- Replaces DROP statements with DROP IF EXISTS for supported postdata reltypes
- Temporarily adds and uses a function that mimics the behavior of DROP CONSTRAINT IF EXISTS, since this feature is not available in Postgres 8.2

Authors: Chris Hajas, Larry Hamel, Stephen Wu