Initialize loop control to handle Oid collision
In 88f0623 a cache of recently used relfilenode Oids was added, and
the cache is interrogated in the GetNewRelFileNode() logic. If the
first Oid in the loop is encountered in the cache however, the loop
will only retry if the compiler has initialized the loop control
variable to true. Explicitly initialize to true to handle this case
for compilers that initialize to false such as clang.