Revert "Make UAO catalog tests deterministic."
This reverts commit 506032ac4a411c4b3da6a16abb1f63ff15e04de8.  VACUUM FULL on
AO/CO is fixed to not give up during drop phase and the tests need not worry
about segfiles in AWAITING_DROP state.

Note that the tests should NOT be run concurrently with any other tests that
use serializable transactions.  VACUUM FULL on AO/CO continues to avoid
deleting segfiles in AWAITING_DROP state in presence of concurrent serializable
transactions that may need the segfiles.