Fix gpload string escape syntax
To match behavior changes of this:

	commit 291f737ab9074b0a89633aaf69468110e0559334
	Author: Andreas Scherbaum <ascherbaum@pivotal.io>
	Date:   Sat May 7 00:08:39 2016 +0200

	    Change the default value of standard_conforming_strings to on

	    This change should be publicized to driver maintainers at once and
	    release-noted as an incompatibility with previous releases.

	    (cherry picked from commit 0839f312e92c7ab0aecb2c4133197a7da7c5fc39)

	    Changes to doc/src/sgml/{config,syntax}.sgml ignored as they are not
	    present in gpdb.

Backslash '\' is no longer recognized as esacpe by default, gpload.py
use quote("\\") as escape string, which translates into two backslashes,
if standard_conforming_strings is on, scanner will recognize it as
double backslashes rather than a single one.

We change escape string to a single backslash to avoid this problem,
which fixes gpload related CCLE cases.

Signed-off-by: Peifeng Qiu <pqiu@pivotal.io>