Fix bug with small sorts and shared scans, introduced by 8.3 merge.
The old GPDB code had a "fall through" in switch-case block, that was not
present in PostgreSQL 8.2. PostgreSQL 8.3 added a new case-block to the
statement, so that the GPDB-specific "fall through" fell through to the
wrong case-block. This lead to a crash, if a sort was small enough to be
performed in memory, was shared by multiple slices. Fix by moving the
GPDB-specific check outside the switch-case statement.