Refactor BFZ stuff to not rely depend on gfile.
This greatly simplifies the error handling and resource cleanup of Bfz
files. Now closing and unlinking bfz files on error is handled by the
resource owner mechanism. This brings the BufFile and BFZ codepaths of
so-called "workfiles", in execWorkFile.c, much closer to each other. Both
now rely on the same mechanism for resource cleanup.

This fixes a crash in the test case that's memorized in the included
regression test.

Remove silly test on bfz_scan_begin(). It was broken, because it used to
set the 'fd' field to 1, which means stdout. That trick doesn't work with
the File interface, it will throw an assertion failure. Rather than fix it,
remove the test altogether. It only tested things that are very explicitly,
directly and unconditionally set in bfz_scan_begin(). There's no real value
in that. Fix the compress_zlib mock test to reflect that the bfz is now
allocated in CurrentMemoryContext rather than TopMemoryContext.

Analysis and test case by Karthikeyan Jambu Rajaraman.