Fix off-by-one bug in quote_literal_internal.
It turns out that PostgreSQL had effectively the same bug, so reported
and fixed (commit 4f5182e1) there too. See:

https://www.postgresql.org/message-id/20161216105001.13334.42819%40wrigleys.postgresql.org

This needs to be refactored, by latest when we merge with PostgreSQL 9.1,
as there is a function called quote_literal_internal there with different
arguments, and the equivalend of GPDB's quote_literal_internal() is called
quote_literal_cstr() there. But for now, just fix the bug.

Fixes github issue #1301, reported by Tang Pengzhou.