[#130928821] Adding CDistributionSpecExternal for external table
This will enable ORCA add RandomMotion or RedistributeMotion on top of
the CPhysicalExternalScan.

That extra motion will redistribute rows across all segments hence
improve the performance of parallel loading using external table.

This behavior is consistent with Planner behavior.

If it is a Master Only table then there is no extra motion added.

Bump the ORCA version to 696.

Signed-off-by: Xin Zhang <xzhang@pivotal.io>