Fix bug with missing Flow on a Result node that replaces an empty Append.
If you perform an DELETE or UPDATE on an inherited table, but the qual
rules out all inheritance children, the planner generates a dummy Result
node with a false One-Time Filter, instead of an Append node. That dummy
Result node did not have a Flow attached to it, which resulted in an
assertion failure later. To fix, attach a Flow to the Result node, like we
do to an Append node in the non-degenerate case.

I wasn't exactly sure what kind of a Flow is appropriate for a dummy node
like this, but mark_plan_general(), i.e. FLOW_SINGLETON, seems to work.
That's what we also use for the similar Result node that we generate for
the case of a HAVING without aggregates or GROUP BY, in grouping_planner().

Fixes github issue #1433.