Backport pg_upgrade from PostgreSQL 9.0.23.
This commit just brings in the upstream code for pg_upgrade and
pg_upgrade_support as is, with no changes. Subsequent work will change
it to work for Greenplum. The goal is to make it usable for upgrading
from Greenplum 4.3 to 5.0.

pg_upgrade_support is a crucial part of pg_upgrade, but lives in a
separate directory to make the build easier.