Add support to pg_dump binary-upgrade for Greenplum
The binary-upgrade mode in pg_dump which was backported in a previous commit
is here extended to handle Greenplum clusters. The gist of the changes are
to extend the Oid dumping to cover all object types as upstream PostgreSQL
only covers types. The changes include:

Make "pg_dump --binary-upgrade" connect in utility mode, it's supposed to
be run against a single segment server. Also make pg_dump work against 4.3
clusters as the binary-upgrade mode will be used with pg_upgrade on the old
cluster in the upgrade.

Dump Oids for all object types in binary upgrade mode. In PostgreSQL only
type Oids are dumped in binary upgrade mode, for Greenplum however we need
to dump the Oids for all objects that require Oid dispatching (see
oid_dispatch.c for details). Inject calls in the dump file for the Oid
dispatcher to preassign Oids during restore.

For partitioned tables, the Oids for each child table as well as their
constraints are added to the dump file. The distributioncheck for
partitions is skipped when restoring during binary upgrade. The
distribution info is not available in the segments so during binary
upgrade we cant enforce this check.

Avoid creating array type for AO relation types, it's not useful for anything
and only grows the catalog for no use.

Heikki Linnakangas, Daniel Gustafsson and Dave Cramer