Fix pg_locks test to not not use the transaction column.
In going from PG 8.2 to 8.3, the pg_locks table does not have a transaction
column anymore.  Instead, it has a virtualtransaction column which is a
composite of the backend number plus some sequential number.

Since the test is trying to find orphan locks, a better test is to look for
locks with NULL owner pids or 0 mppsessionid.

The type of current_setting('gp_session_id') is also changed from integer to
text.  Therefore, this is also casted to int.