Remove test_skip_checkpoint_abort_transaction.py test.
The test_skip_checkpoint_abort_transaction.py test is testing fix for
MPP-17297/MPP-17817 where a failure in commit processing the standby was not
handled properly by the master.  However, the code changes in the original bug
fix in 4.2.3.0 are no longer applicable to the 5.0 code line, and the replication
mechanism between standy and master has changed to walrep since then.

Therefore, I am removing this test altogether.