Update Version number macros
We bumped the GPORCA version to 2.0 after merging GPOS with GPORCA.
Cmake undefs the GPORCA_VERSION_MINOR when the minor number is 0. 0 is
considered as a false constant by cmake.

In gpdb, use GPORCA_VERSION_STRING instead of separate GPORCA_VERSION_MAJOR
and GPORCA_VERSION_MINOR.

Bumped the GPORCA version to 2.1

Signed-off-by: Dhanashree Kashid <dkashid@pivotal.io>