Move duplicate handling of LOCATION to the parsing stage
Creating an external table with duplicate location URIs is a syntax
error so move the duplicate check to the parsing step. This simplifies
the external table command code as well as provides an error message
with context marker to the user.

Also fix up test sources to match the new error messages and remove
the duplicate test from gpfdist since we already test that in ICW.