TINC add wait_for_db for resilience to failover_to_mirror tests
Seen failures multiple times in pipeline at point for checking if database is in
changetracking or not for failover_to_mirror scenarios. For this case primary is
paniced and mirror takes over but due to work-load running, master may see
panics due to 2PC erroring out after retries. The test should be able to
continue working in such a condition, hence adding wait_for_database to be up
after master panic.