Fix dispatching of pg_constraint OID assignments.
Constraints are not uniquely identified by namespace and name alone.
It's possible to have multiple constraints with the same name, but attached
to different tables or domains. So include parent relation and domain OIDs
in the key.

This was uncovered by a regression test in the TINC test suite. Move the
OID consistency tests from TINC into the normal regression test suite, to
have test coverage for this.