Full vacuum on heap table should not commit the transaction in `repair_frag()`.
In GPDB, vacuum full on heap relations is divided into two steps. The first
step moves the tuples of the relation to free the last pages. The second step
truncates the relation. Both the steps are performed in two separate
transactions so we don't need to commit the transaction in `repair_frag()`.