Fix dispatch of pre-assigned OIDs in queries with initplans.
If a CTAS query uses a query with an InitPlan, and the InitPlan contains
a call to a function that dispatches another query to the segments, the
list of Oids is dispatched to the segments prematurely, in the inner query.
As a result, when the CTAS query is dispatched, the OID of the new table
is not included, and you get a "no pre-assigned OID for relation" error.

To fix, attach the OID of the new table to the query descriptor before
executing the InitPlans.

Fixes github issue #1543, reported with test case by Abhijit Subramanya.