Properly handle unlink failure for RemoveOldXlogFiles
Originally, the return code was ignored. Now, if unlink failed for any
reason, the error message will be logged, and also adjust the
CheckpointStats properly.

Now MirroredFlatFile_Drop() returned the same return code as unlink(),
and also the errno is preserved, so the correct error message can be
logged by the caller.

Signed-off-by: Xin Zhang <xzhang@pivotal.io>