Attempt to make LICENSE discoverable by Github
Github auto discovers the project license via the licensee ruby gem
which in turn bases it's matching on the choosealicense.com database
of license texts. The matching is failing for our license so copy
the exact text from choosealicense in an attempt to make the matching
work. The diff pulls in the Appendix and changes whitespace on the
license name line, no changes whatsoever are made on the main license
text.