Refactor binary-upgrade support in pg_dump
The binary-upgrade support was added with the pg_upgrade patch and
allows for the Oids in the old cluster to be pre-assigned to their
respective objects during pg_upgrade. The main motivation for this
refactoring is that we need to pre-assign all Oids before object
creation during restore into the new cluster since a restore can
otherwise allocate an Oid which is later preassigned. This brings
all pre-assignments into the head of the dumpfile such that they
happen before most object creations (pg_dumpall objects are still
written to the dumpfile before pg_dump has a chance).

The main contributions of this patch include:

* All binary-upgrade methods are moved to binary_upgrade.{c|h} for
  pg_dump and binary_upgradeall.{c[h} for pg_dumpall. This greatly
  reduce the diff wrt upstream PostgreSQL
* Oid preassign calls are now loaded into the backup archive as
  TOC entries which allows for them to be sorted to be output
  before object creation
* Avoid usage of PQExpBuffers for trivial string construction where
  a fixed buffer on the stack is sufficient, or simply outputting
  the string directly
* Binary upgrade dumping in pg_dump is moved to a separate function
  which loops over the dumpable objects instead of being mixed in
  with the main code
* Various simplifications of the code and cleanups where possible