Remove unnecessary ORCA-related cruft from makefiles.
The .so files in the ORCA release tarballs are directly in a "lib"
directory, so the OBJDIR_DEFAULT stuff, to search in a directory like
".obj.linux-x86_64.opt", is obsolete.

The XERCES, OPTIMIZER et al. environment variables were not set by
anything. We rely on the --with-libs and --with-includes to be set
correctly in the configure command line, so remove the env variable stuff.