Add config.h, for options that affect binary compatibility.
Before this patch, consumers of ORCA had to know out-of-band which
flags were used to compile ORCA, because e.g. ORCA compiled with
GPOS_DEBUG would only work if the application using ORCA was also
compiled with GPOS_DEBUG. This is because many of the structs differ
depending on GPOS_DEBUG. Same for the architecture flags, like GPOS_i386.

The new config.h file is #included from a few central other header files,
to make sure it gets included in any application that uses other gpos
headers. We probably should include config.h from all other gpos header
files, to be sure, but this seems to be enough for ORCA itself and GPDB at
least.

Bump version number to 2.2.