Silence compiler warnings when building without codegen.
I'm seeing these warnings:

execProcnode.c: In function ‘ExecInitNode’:
../../../src/include/codegen/codegen_wrapper.h:56:62: warning: statement with no effect [-Wunused-value]
 #define CodeGeneratorManagerAccumulateExplainString(manager) 1
                                                              ^
execProcnode.c:803:5: note: in expansion of macro ‘CodeGeneratorManagerAccumulateExplainString’
     CodeGeneratorManagerAccumulateExplainString(CodegenManager);
     ^
../../../src/include/codegen/codegen_wrapper.h:54:64: warning: statement with no effect [-Wunused-value]
 #define CodeGeneratorManagerPrepareGeneratedFunctions(manager) 1
                                                                ^
execProcnode.c:807:5: note: in expansion of macro ‘CodeGeneratorManagerPrepareGeneratedFunctions’
     CodeGeneratorManagerPrepareGeneratedFunctions(CodegenManager);
     ^

To fix, define each dummy macro to return a value that's consistent with
the actual function. For functions returning void, use "((void) 1)" to hint
the compiler that it's OK to ignore the result. The void functions is what
the warnings actually complained about, but fix them all for tidyness.