Disable MyXactAccessedTempRel
In Postgres, PREPARE TRANSACTION errors out if it accessed a temp relation.

In GPDB, two phase commit is always used, even for transactions accessing temp
relations, which is used by MPP query processing.

In this case, GPDB treats temp table as regular distributed table, but shorten
its life cycle to sessions or transactions.

GPDB also stores temp table information in shared buffers, hence MyXactAccessedTempRel
is not used, and we are disabling it with comments. The reason to keep the code around
is to help future Postgres upstream merge.

Signed-off-by: Xin Zhang <xzhang@pivotal.io>