Make gpinitstandby use fast checkpoint during pg_basebackup
The 8.3 merge brought in a change to BufferSync() introducing the
concept of a slow checkpoint which throttles I/O so that regular
background checkpoints do not heavily affect other operations. Our
pg_basebackup call in gpinitstandby was not changed to bypass this new
functionality so here we make it use fast checkpoint to do so. The
fast arg was not utilized yet to send CHECKPOINT_IMMEDIATE to
RequestCheckpoint() so brought in the change from upstream Postgres.

Referenced upstream Postgres commit:
https://github.com/postgres/postgres/commit/387060951e40d550d37fe0457521e900d8c60feb