Append `.0` to version string
This patch adds a `.0` (or as they call it over semver.org, the "patch"
number) to our version. This makes our version semver-compliant. It also
makes it abundantly clear that our version number (say `2.2.0`) is not a
decimal floating point number. ("What is newer? 2.2 or 2.19?")

N.B. our practice is only bumping the minor and major version numbers.
This change doesn't not signal a change in that practice.

I'm also hiding this commit from CI so that it only impacts the next
push (so it will be, say, `2.3.0` if we are currently on `2.2`).

[ci skip]