Fix disabling of -Werror in PGAC_PROG_CXX_CXXFLAGS_OPT macro
This was copy-pasted from PGAC_PROG_CC_CFLAGS_OPT, but for the g++
compiler, we need to set ac_cxx_werror rather than ac_c_werror.

The point of this dance with Werror is to detect if the compiler accepts a
flags, but causes warnings, like:

cc1plus: warning: command line option ‘-Wmissing-prototypes’ is valid for C/ObjC but not for C++

We don't want to use such a flag.