Add upload to Coverity step and set twice weekly trigger to pipeline [#135626599]
We fixed all the shell checks for the scan_with_coverity script and now
the cov-build is run in gpAux. We added the following step to tar it and upload
the tarball to scan.coverity.com.

In the pipeline file we added a twice weekly trigger on Monday and
Thursday morning.

Signed-off-by: Jingyi Mei <jmei@pivotal.io>
Signed-off-by: Tom Meyer <tmeyer@pivotal.io>