Remove gppkg packaging of PL/perl.
In previous versions of GPDB, we compiled PL/perl against the Perl version
that shipped with RHEL5, and built a separate gpkkg package of plperl,
compiled against the version that ships with RHEL6. If you wanted to use
PL/perl on RHEL6, you had to install the package separately. That's now
obsolete, as we don't support RHEL5 anymore. We can just always build
against the later Perl version.

Adjust concourse script to cope when no gppkg packages were built, as is
now the case on most platforms.