Move patterns used only by a particular test out of global init_file.
This reduces the risk of accidentally masking out messages in a test that's
not supposed to produce such messages in the first place, and is just
nicer in general, IMHO.

While we're at it, add a brief comment to init_file to explain what it's
for. Also, remove a few more matchsubs from atmsort.pm that seem to be
unused.