Remove StreamBitmap-related dead code
If GPDB evaluates a query that needs to access an index, e.g., btree, then at
least one Bitmap Index Scan will appear in query's plan.
MultiExecBitmapIndexScan function will be invoked to construct the bitmap. If a
query contains a WHERE clause similar to d in (1,2), where a bitmap index has
been created on column d), then instead of creating a simple bitmap,
MultiExecBitmapIndexScan will generate a composite bitmap that ORs all the
bitmaps that satisfy the query condition. In our example,
MultiExecBitmapIndexScan (in particular bmgetmulti) will OR the bitmaps of d =
1 and d = 2.

To OR two bitmaps A and B, A should be a StreamBitmap (and not a HashBitmap).

In this commit, we remove code that was executed when MultiExecBitmapIndexScan has
to generate a composite bitmap of A and B, where A is a StreamBitmap and B is a
HashBitmap generated when accessing either a btree or a gin or a gist or a hash
index. It seems that this is not a possible scenario because i) A can be a
StreamBitmap only if a bitmap index has been constructed on a particular
column, and ii) to the best of our knowledge we cannot have a single Bitmap
Index Scan that will access both a Bitmap index and an index of another type
e.g., btree.

Authors: Nikos Armenatzoglou
         Shreedhar Hardikar <shardikar@pivotal.io>
         Haisheng Yuan <hyuan@pivotal.io>