Take upstream lazy_truncate_heap locking behavior
During lazy_truncate_heap, an exclusive lock is taken on the heap to be
truncated. This lock is required to prevent other concurrent transactions
reading an invalid rd_targblock which is going to be propogated to other
backends as part of cache invalidation at commit time.

This lock need to be held till the end of commit, hence we remove the
UnlockRelation() added for GPDB, which is introduced to avoid a deadlock
situation caused by concurrent vacuums.

However, we cannot repro this deadlock on latest GPDB, where this deadlock was
found in a very early version of GPDB (back to 3.3).

Signed-off-by: Xin Zhang <xzhang@pivotal.io>