Add GUC gp_allow_rename_relation_without_lock for renamerel().
Originally, we don't hold AccessExclusiveLock while renaming relations to provide
supportability for altering relation name while directly modifying system catalogs.

However, this breaks the isolation semantics of ALTER TABLE. Hence, we add the lock
to make it consistent with upstream behavior.  The original behavior is now
possible with the GUC.

Signed-off-by: Xin Zhang <xzhang@pivotal.io>