Update string escaping to account for standard_conforming_strings.
Because the standard_conforming_strings GUC has been updated to be
on by default, using pg.escape_string (which assumes that the GUC
is off) is no longer sufficient to escape all strings in backup
and restore.  Backup and restore now use pg.DB().escape_string,
which connects to the database so that it can use the correct value
for standard_conforming_strings.

Authors: Karen Huddleston, Jamie McAtamney