Fix memory leak and error handling of workfiles.
There's no need to allocate ExecWorkFile struct in TopMemoryContext. We
don't rely on a callback to clean them up anymore, they will be closed
by the normal ResourceOwner and MemoryContext mechanisms. Otherwise,
the memory allocated for the struct is leaked on abort.

In the passing, make the error handling more robust.

Also modify tuplestorenew.c, to allocate everything in a MemoryContext,
rather than using gp_malloc() directly. ResourceOwner and MemoryContext
mechanisms will now take care of cleanup on abort, instead of the custom
end-of-xact callback.

Remove execWorkFile mock test. The only thing it tested was that
ExecWorkFile was allocated in TopMemoryContext. That didn't seem very
interesting to begin with, and the assumption is now false, anyway. We
could've changed the test to test that it's allocated in
CurrentMemoryContext instead, but I don't see the point.

Per report in private from Karthikeyan Rajaraman. Test cases also by him,
with some cleanup by me.