Removing the FIX_ME comment for assert on operator memory.
Due to the plan caching, we might use the same plan again. In that case, operatormem is not 0, to begin with, and it will have previous values. So this assert assumption is wrong and hence removing it. Based on the investigation we found that,

1. Operator Memory is being initialized even if statement_mem get
changed when we use the cached plan.
2. We also confirmed that even though we use plan cache, we will still call the executor init.
3. MemoryAccountId may be another potential candidate in Plan to have a
similar issue. But w.r.t memory account id, we confirmed that is not
serialized from QD to QE and they all get reinitialized anyway when we
do the ExecInit.
Note: We don't serialize memory account id because it is an index to the
account array specific to the current process only.