Backport the pg_upgrade-related changes from upstream commit 3d21a0f300.
This patch was backported earlier already, but now that we cherry-picked
the binary-upgrade work from 9.0, we must also pick the related parts
of 3d21a0f300, that were not applicable earlier.

In the passing, also pick the "if (secure)" checks to EchoQuery and
UseNewLine from the upstream commit. I don't know why those were not
picked earlier.