Fix TINC aoco_compression tests
The storage type for cidr and inet was changed from plain to main
around the beginning of open source Greenplum but still plagues some
tests. These should be the last ones inside the TINC directory. Other
than that, just some random ans file updates.

This commit is just to make the tests green on CI. The tests will be
refactored in a later commit to run more quickly and to reduce
clutter.