Port Bugbuster Oid consistency test to ICW
Almost all of the tests performed in the Bugbuster Oid consistency
test were already present in the ICW counterpart, the remaining few
of interest were pulled over (one on pg_trigger and some more on
pg_constraint).

The Bugbuster test did however not actually test Oid consistency at
all since the test result was never matched to an expected output.
Instead the output was a count of the number of ERRORs encountered
during the test run, all of which were bogus errors from setting a
GUC that was long since removed.