Clean up url_* API.
The callers of url_fclose() didn't check the return value, and most of the
implementations didn't return anything interesting anyway. Mark it as
void. The implementations are expected to elog() an error instead, if
something goes wrong.

Merge the 'size' and 'nmemb' parameters to url_fread and url_write to just
one 'size' parameter. There's no need to precisely emulate the fread/fwrite
functions here. The callers just always passed size == 1, and there was
even an Assert(size == 1) in url_file_fread() implementation. The
implementations also differed on how the return value was calculated.