Fixed gpversion and gpmigrator_util tests
This commit expands on 2ccdfbf25dc991632659a3c5f8519f65c5091d16, as
the gpversion helper functions also needed to be modified to account
for the 4-digits-to-3-digits conversion; with 3 digits, the major
version is 5, not 5.0.

It also updates the gpmigrator_util tests to use 4.3 as the previous
version, instead of 4.2, and removes some references to customers in
the test version strings.