Fix incorrect size of sliceMap in CdbDispatchResults. (#1701)
* Fix incorrect size of sliceMap in CdbDispatchResults.

In case the slice table of a plan contains multiple roots, we created a
sliceMap based on the slice capacity of only one of trees. Since sliceIndices
reference the entire slice table, we ended up referencing sliceMap with
sliceIndex out of bounds of the sliceCapacity. This fixes that issue by
correctly calculating sliceCapacity.
Refer commits that introduced this problem :
 [1] 4b3609423ae8e1d269bf0b02238fd05af7ca495d PR#827
 [2] a2ecd1fa97e8e517725c6fd9a10f06f53944fd34

[1] Combined the calculations of resultCapcity & sliceCapcity, which is
    incorrect - e.g in cases that have an InitPlan
[2] Removed resultCapacity considering it redundant.

* Revert "Fix incorrect size of sliceMap in CdbDispatchResults."

This reverts commit 8cae828809ad22aaf05ebfd77f4be35fe3614e63.

* Update small changes to fix the issue