Enable parallel make in the pipeline.
There was a mechanism, using PARALLEL_BUILD=1, to guess suitable -j and -l
flags to pass to make. I tried enabling them, and it came up with options
"-l3.5 -j", which doesn't seem right, and didn't speed up the build. Instead
of trying to be smart, let's just hard-code "-j4" into the concourse
launcher script. That's parallel enough to speed up the build considerably,
but also not too parallel, so that if the system is really underpowered,
it should still not harm too much.