Remove uninteresting test objects and case.
These tables and views were not used for any further testing. They are
uninteresting in themselves, we have a lot more interesting tests
elsewhere for tables, views, and comments on them.

In the test for upper case tables, one INSERT is enough to confirm that it
works.

We have more comprehensive testing of transactions in the 'transactions',
and elsewhere. The 'employee' case is quite unremarkable, so just remove
it.