Backport OpenTransientFile from PostgreSQL 9.3.
OpenTransientFile() opens an arbitrary file, like open(), but will track
the file descriptor by the resource owner mechanism, so that it is
automatically closed on abort. We already had a similar mechanism for
temporary files, and for files opened with fopen(), this fills the gap of
when you want a plain file descriptor rather than a FILE *.

I need this for the next commit, which will use this to automatically close
files opened by the fstream.c on abort.

This is a backport of upstream commit 1f67078ea32. I only backported the
addition of the new function, not the changes to the various callers to
use the new function, to minimize merge conflicts.