Concourse pipeline remove commit trigger on cs-aoco-compression job.
The cs-aoco-compression job was added with intention to run on a CRON
instead of on every commit batch. However, I didn't notice that the
aggregate it points to contains a commit trigger... so the project
needs its own aggregate. This aggregate can be turned into a reference
later if any other test jobs become CRON triggered.