Remove inapplicable sanity check [#121649701]
`rowMarks` in a Query object represent relations whose tuples are subject
to locking when a query contains `FOR SHARE` or `FOR UPDATE` clauses in
a SELECT. Those rowMarks are wholesale carried over to an EState, via
PlannedStmt.

In Postgres the optimizer (planner) would add a junk TID column for each
table in the rowMarks. And the executor will lock the tuples when
executing the root node in the plan.

The check we are removing seems dormant even in Postgres: to actually
cover the `elog(ERROR)`, the code path needs to have
  0. No junk TID columns
  0. A non-empty rowMarks list

In Postgres the above conditions are all-or-nothing. So the check feels
more like an `Assert` rather than an `elog(ERROR)`, and it seems to be
dead code.

In Greenplum, the above condition does not hold. Greenplum has a
different (but compliant) implementation, where we place a lock on each
table in the rowMarks.  Consequently, we do *not* generate the junk
column in the plan (except for tables that are only on the master). This
is reasonable because the root node is most likely executing in a
different slice from the leaf nodes (which are most likely SCAN's): by
the time a (hypothetical) TID reaches the root node, there's no
guarantee that the tuple has not passed through a motion node, hence it
makes no sense locking it.

This commit removes that check.