Rewrite readindex test cases so that they work with ORCA.
The "CREATE TABLE ... AS SELECT ..." commands that were used to create
the test tables choose random distribution policy when ORCA is used, while
without ORCA, it chooses the first column as the distribution key. I'm not
sure why, and it seems inconsistent, but work around that by using CREATE
TABLE + INSERT ... SELECT to create and populate the tables instead.

With CTAS, the all the rows in AO tables go to AO segfile 0, while with
CREATE TABLE + separate INSERT, they go to segfile 1, so update the
expected output for that.