Remove tests for --help.
These tests merely test that "gpinitactivestandby --help" and
"gpinitstandby --help" produce a particular output. That's not very
interesting, --help seems highly unlikely to suddenly stop working. Also,
memorizing the exact help output in a test adds a hurdle to keeping the
--help text up-to-date - you will need to remember to update the
test too, whenever you update the help text.

What would be useful, is a test to check that --help covers all the options
and provides useful advice to a DBA. But that cannot be automated.