Move and fix test for uniqueness checks on RENAME PARTITION.
The old test case was no longer testing what it was supposed to. The
ancient bug reported as MPP-3115 was related to renaming a partition,
but the behavior has changed since then, to not allow two partitions to
have the same name.

I couldn't see any tests that RENAME PARTITION enforces uniqueness, except
for the special case that the new name is the same as the old, so I added
a test for that to partition1.sql. Other than that, I think we can simply
remove this test case as obsolete.