Remove uninteresting tests on subpartition syntax.
Subpartition names are SQL identifiers, so you cannot use a plain integer
as name, unless you quote it. There isn't really anything special here,
and never was. The original bug report was:

> I was trying to add subpartitions named 2001, 2002, 2003, etc. but was not
> allowed:
> ...
> However it was allowed when the name was changed from a number to a string

It was closed as "No fix needed", as that is expected behavior. I'm not sure
why we bothered to add test cases for this originally, but I think we can
just drop them now.