Remove uninteresting UAO fault-injection test cases.
The pattern in these tests was:

1. Create and populate a test append-only table
2. Induce a fault with gpfaultinjector, to crash the server at beginning of
   appendonly_insert/update/delete function.
3. Try to INSERT, UPDATE or DELETE some rows, crashing the server.

Crashing the server at the beginning of the operation isn't very
interesting, because we haven't done any on-disk modifications at that point
yet. There is no reason to believe that there would be a problem at crash
recovery at that point. In particular, after the crash, the state at disk
will look identical in the INSERT, UPDATE and DELETE scenarios.

Therefore, remove the tests.