Fix test VACUUM FREEZE case, and try to make it more robust.
I typoed the INSERT command in this test, and even checked in the ERROR in
the expected output. Fix.

Also, make test case a bit less fragile, by advancing the XID counter also
after the INSERT+DELETEs in the second part of these tests. This ensures
that the tables will look 'old' before the vacuum.

Tweak the limits of 'old' and 'young' XID, to make more sure that a table
just created is considered 'young', and that advancing the XID counter
pushes the age to 'old' on all segments.