Remove obsolete PL/perl test.
These test queries were copy-pasted from the upstream test at
src/pl/plperl/plperl_trigger.sql, so nothing interesting there.

The test harness would try to install plperl with gppkg, but I don't think
that's very interesting to test. AFAIK we don't plan to install plperl.so
with the gppkg mechanism anymore. And if we do, the tests for that should
live next to the gppkg scripts to create the package, not here.

test_spi.py referenced query01.sql, query02.sql and query03.sql, but the
directory only contained query01.sql and query03.sql. Therefore, I believe
this was actually broken as it is.