Remove uninteresting uao_load tests.
We have sufficient test coverage for gpload in the Behave tests in
gpMgmt/test/behave/mgmt_utils. And we have plenty of tests for external
tables elsewhere.

The combination of using an external table or gpload as the source, and
append-only table as the target, is not particularly interesting. So remove
the tests.