Remove tests on DROP DATABASE while the database is in use.
These tests are all variations:

1. Open a session to database, do some operations in it like CREATE INDEX,
   DROP INDEX or REINDEX.
2. In another session, attempt to drop the database.

The result is always that dropping the database fails, after a timeout,
with error: ERROR:  database "reindexdb" is being accessed by other users.

That's not a very interesting test case, because you get that same error
regardless of what the other backend is doing. It doesn't even need to
be doing anything, even an idle session has the same effect. Hence, remove
these pointless tests.