Remove redundant gp_default_storage_options test from TINC.
The TINC test tested that if you set gp_default_storage_options to
'checksum=False' and create tables, the tables are created without
checksums. We already have tests for that in the 'dsp' test in the main
test suite.

Add the 'checksum' column to a couple of validation queries in the 'dsp'
test, to ensure that it would catch it if setting checksums on/off didn't
work.