Remove bogus checks that REINDEX chooses same relfilenode on all segments.
REINDEX chooses the new relfilenode for an index independently in each
segment (see also 88f0623ea1). Therefore, it is not safe to assume, that
after REINDEX, the relfilenode would be the same in all segments, if there
is concurrent activity. These tests failed a few times in the concourse
pipeline. The purpose of this test, per the @description, is to test
permissions, so I'm not sure why these relfilenode checks were tacked on to
it in the first place.