Move gp_default_storage_options related test from TINC.
The 'add_part_checksum_with_guc_enable' test tested that a newly-created
table has checksums enabled, when you don't set gp_default_storage_options,
and you explicitly set checksums on in the WITH clause. That's not very
interesting, because checksums default to 'on' anyway. We have plenty of
coverage of that default already.

The 'add_part_checksum' test was marginally useful. It tested that
if you set checksums off in the GUC, but explicitly specify checksums=on
in the CREATE TABLE, checksums really are enabled. And it tested the same
for ALTER TABLE ADD PARTITION. Move that test to the 'dsp' test in the
main regression suite.