Remove unnecessary DROP + CREATE DATABASE from tests.
The 'gptest' database was actually used by the tests. CREATE DATABASE is
a fairly expensive operation, so let's avoid doing it without a good
reason.

I don't actually have means to test this properly, so I'm going to push
this in blind and see what Concourse thinks of it. If something goes awry,
I'll revert.