Remove deprecated memory control GUCs
The gp_eager_hashtable_release GUC was deprecated in version 4.2
in 2011 when the generic eager free framework was implemented. The
leftover gp_eager_hashtable_release was asserted to be true and
never intended to be turned off. The same body of work deprecated
the max_work_mem setting which was bounding the work_mem setting.
While not technically tied to eager hashtable release, remove as
well since it's deprecated, undocumented and not terribly useful.

Relevant commit in closed source repo is 88986b7d