Remove extended ALTER TABLE INHERIT syntax
The gp_enable_alter_table_inherit_cols GUC was used to allow a list
of columns to override the attribute discovery for inheritance in
ALTER TABLE INHERIT. According to code comments the only consumer
of this was gpmigrator, but no callsite remains and no support in
pg_dumpall remains either. Remove the leftovers of this to avoid a
potential footgun and get us closer to upstream code.