Remove redundant tests on RLE delta compression.
There were three variants of the same test, combined with 4 compresslevels.
The "*_null" variants insert rows, such that the delta between some rows is
positive, and between others, negative. And it also contains some NULLs.
The other variants only contained positive or postive and negative deltas,
without NULLs. It doesn't seem interesting to have separate tests for the
more narrow cases. If there's something wrong with the handling of positive
or negative deltas, the "*_nulls" tests will catch that too.