Remove "queryfinish" test, because it doesn't add any meaningful coverage.
The test launched one backend, and ran ALTER TABLE ADD COLUMN in it. Before
committing the ALTER TABLE transaction, it launched another backend, and
ran a query against the table. That will block, because the ALTER TABLE
is holding an AccessExclusiveLock on the table. It then committed the
first transaction, letting the second transaction to continue.

That's not a very interesting test case. ALTER TABLE grabs an
AccessExclusiveLock very early, so it's hard to imagine it would fail to
hold it. Also hard to imagine what could go wrong in having the other query
block on the lock. Backends frequently block on all kinds of locks, for
short periods of time, so just blocking on a lock isn't very interesting.
Furthermore, AFAICS, if the ALTER TABLE somehow failed to acquire the lock,
so that the query didn't block, the test would return the same output, and
therefore wouldn't catch that case.