Implement CREATE RESOURCE GROUP syntax.
There are two default resource groups 'default_group' and 'admin_group',
to create more please use the CREATE RESOURCE GROUP command, group
options can be specified with the WITH clause, at the moment
'cpu_rate_limit' and 'memory_limit' are mandatory, other options are
all optional.

    -- create a resource group
    CREATE RESOURCE GROUP rg1 WITH (
        concurrency=1,
        cpu_rate_limit=.2,
        memory_limit=.2
    );
    -- query the resource group
    SELECT oid FROM pg_resgroup WHERE rsgname='rg1';
    SELECT * from gp_toolkit.gp_resgroup_config
        WHERE groupname='rg1';
    -- create/alter a role and assign it to this group
    CREATE ROLE r1 RESOURCE GROUP rg1;
    ALTER ROLE r2 RESOURCE GROUP rg1;

*NOTE*: this commit only implement the SQL syntax, the actual resource
limitation will not take effect at the moment as the resource group is
still under development.

*NOTE*: test cases are not included in this commit as once a testing
resgroup is created it can't be dropped due to lack of DROP syntax, so
the test case can't be re-run and will introduce side-effect to the
system. So it's better to provide test cases after the DROP RESOURCE
GROUP is implemented.

Signed-off-by: Ning Yu <nyu@pivotal.io>