Prevent deadlock between a committing backend and checkpointer process.
The deadlock occurs when a backend is about to unlink a file but finds the
shared fsync requests queue full. If the backend waits for checkpointer to
consume fsync requests from the queue, there is a deadlock. Because
checkpointer is already waiting for this backend on `MyProc->inCommit`.

A backend that is about to unlink a file is going through commit processing and
must already set `MyProc->inCommit`.  This commit prevents the deadlock by
checkpoint process absorbing requests from the shared queue while waiting for
the transactions to commit.

Fixes issue discussed on mailing list at:
https://groups.google.com/a/greenplum.org/forum/#!topic/gpdb-dev/PHKuQPNwWs0