Fix planning of NOT IN (<empty set>).
The transformation of "NOT IN" or "= ALL" incorrectly added a condition
that the outer side of the join must not be NULL. But that's not true,
when the other side is an empty set.

Discovered while moving old TINC tests to the main test suite. The
'lasj_hash_all_3' test exposed this issue, when run with ORCA disabled.
This has apparently gone unnoticed, because no-one has run those tests
without ORCA. There were also incorrect query results memorized, for tests
'lasj_notin_multiple_3' and 'lasj_hash_notin_multiple_3'. They are similar
queries, but because ORCA falls back to the planner for them, we had
memorized the Planner results for them, and hadn't noticed that the
results were in fact incorrect.

Fixes github issue #1907.