Remove tests on REINDEX and TRUNCATE/ALTER/DROP TABLE running in parallel.
REINDEX grabs a ShareLock on the table relation very early on, and
TRUNCATE grabs an AccessExclusiveLock very early on. If we trust that the
lock manager isn't broken, there's no need to test this combination of
commands in particular. (For testing the lock manager, these tests give
only very narrow coverage.) Furthermore, all these test cases are
effectively the same, as it it makes no difference what kind of a table or
what kind of an index it is, the locking works the same for all.

ALTER and DROP TABLE also grab an AccessExclusiveLock on the table, which
conflicts with the ShareLock that REINDEX takes. It works the same with all
flavors of ALTER and DROP TABLE, and with all kinds of tables, and all
kinds of indexes.