Fix test case so that atmsort.pm recognizes EXPLAIN as such.
If the EXPLAIN is on separate line, the mechansim in atmsort.pm to detect
EXPLAIN queries, and to mask out trivial differences like different costs,
doesn't recognize it. This caused a failure, for example if you set
"optimizer=off" explicitly in postgresql.conf, because that added a
Settings line to the explain output. We rely on atmsort.pm to mask out
trivial differences like that.