Fix representation of CONST in explain plan
When `explain` traces a VAR to a CONST, it assumes (rightly so) that
it's most likely the only constant being projected at that operator,
e.g.

```
EXPLAIN SELECT a AS c FROM (SELECT 1 AS a) AS bar ORDER BY 1;
                   QUERY PLAN
------------------------------------------------
 Sort  (cost=0.03..0.04 rows=1 width=4)
   Sort Key: (1)
   ->  Result  (cost=0.00..0.01 rows=1 width=0)
 Optimizer status: legacy query optimizer
(4 rows)

```

For a query containing the `VALUES` construct, planner would generate
`ValuesScan`, whereas ORCA generates an `Append` with each constant as a
`Result` node underneath. Notice how the ORCA plan shape breaks the
assumption of the `EXPLAIN` rendering logic, resulting in somehow
confusing output like this:

```
explain select * from foo, (values (1), (2)) f(i) where foo.a=f.i;
                                  QUERY PLAN
------------------------------------------------------------------------------
 Gather Motion 3:1  (slice1; segments: 3)  (cost=0.00..431.00 rows=1 width=8)
   ->  Hash Join  (cost=0.00..431.00 rows=1 width=8)
         Hash Cond: (1) = a
         ->  Append  (cost=0.00..0.00 rows=1 width=4)
               ->  Result  (cost=0.00..0.00 rows=1 width=4)
                     ->  Result  (cost=0.00..0.00 rows=1 width=4)
               ->  Result  (cost=0.00..0.00 rows=1 width=4)
                     ->  Result  (cost=0.00..0.00 rows=1 width=4)
         ->  Hash  (cost=431.00..431.00 rows=1 width=4)
               ->  Table Scan on foo  (cost=0.00..431.00 rows=1 width=4)
 Settings:  optimizer=on
 Optimizer status: PQO version 2.7.0
(12 rows)
```
In the above case, `(1)` is less than helpful because that is not the
only constant in the relation.

This commit renders a `VAR` node that trace back to a `CONST` that has a
non empty `resname` with the `resname` instead of the constant value.

An unintended (but good) consequence of this change is that some
constants are now rendered as their projected column instead of the
constant value, e.g.:

```
EXPLAIN SELECT a AS c FROM (SELECT 1 AS a) AS bar ORDER BY 1;
                   QUERY PLAN
------------------------------------------------
 Sort  (cost=0.03..0.04 rows=1 width=4)
   Sort Key: c
   ->  Result  (cost=0.00..0.01 rows=1 width=0)
 Optimizer status: legacy query optimizer
(4 rows)
```

Signed-off-by: Jesse Zhang <sbjesse@gmail.com>
Signed-off-by: Bhunvesh Chaudhary <bchaudhary@pivotal.io>