Print regression.diffs for gphdfs on CI
The generated bash scripts for running the regression tests are
executing with "set -exo pipefail" where "-o pipefail" makes the
script exit on the first error. Since pg_regress will exit with
returncode 1 on a run with a failed suite, the script will exit
before reaching the check for regression.diffs. Fix by adding a
trap function instead like the other CI scripts.

This use the same construction as other CI scripts of using find
rather than hardcoding the path, which can be seen as overkill
since there's currently only a single file. This is however a
cheap operation and will ensure that we don't miss any future
regression.diffs added somewhere, or a chang to  pg_regress
outputdir parameter.