Fix broken free_readfile()
Commit 5a2d9a5aad cherry-picked some code to plug a memory leak from
test_postmaster_connection(), but the surrounding code looks quite
different in the upstream version where it was cherry-picked from. In our
8.3-based version of the function, 'optlines' is modified in the loop, and
doesn't point to the original allocated array anymore, so we passed a bogus
pointer to free().

This made e.g. "pg_ctl restart -w" throw GLIBC warnings.

Remove the offending call to free_readfile(), reverting that back to the
way it was before commit 5a2d9a5aad. Yes, that re-introduces the memory
leak, but that's not a big deal. pg_ctl is not a server program, it will
exit as soon as it gets its job done, so the leak is very short-lived.
It was plugged only to silence Coverity warnings in the first place.