Remove more redundant TINC tests on int and int4 types.
In commit 989c502f39, I already removed a bunch of these, but didn't catch
them all then. This removes the remaining tests on "int", where an
identical test on "int4" also exists. "int" is just an alias for "int4", so
there is no difference.

I just realized the reason these tests existed in the first place: the
"int" tests were supposed to actually test the "int2" datatype, not "int4".
The test values in the "int" test have been chosen so that they would fit
in a 16-bit integer. So if we wanted to restore the original purpose of
these tests, we should search & replace "int" with "int2". But testing all
the different datatypes isn't too interesting anyway, and there's no reason
to believe that int2 would be a special, and at the same time not test many
datatypes with different lengths, alignments and pass-by-value flags, so
let's just remove these.