Remove uninteresting tests on using + operator in DML queries.
These expressions are as simple as it gets. If any of these queries would
fail, the system would be so broken that you would notice it very quickly.

(These tests had nothing to do with MPP-21090, like the description tags
claimed. That bug happened when you dropped a column and then added a
partition, and there are no DROP COLUMNS or ADD PARTITIONS in these tests.)