Fix ON clause dumping for external tables
Commit ac2fd68027 introduced ON MASTER for all external tables, a
clause until then only available on EXECUTE. The support in pg_dump
did however assume there was an ON clause, which isn't neccesarily
the case for external tables in 4.3 (and below), causing "illegal
ON clause" logging without setting an appropriate ON clause. Fix
by adding support for non-5.0 servers and rearrange logic to make
it clearer.

Also; error out on incorrect ON clause, as that could indicate
catalog corruption and include the relation name in the error
logging to aid debugging; and use static string/char PQExpBuffer
functions where possible to avoid overhead.

Apply to pg_dump and cdb_dump_agent.