Add function for throwing ereport-like errors from gpopt code.
This allows us to have the exact same error message and hint for errors,
as what the traditional planner produces. That makes testing easier, as
you don't need to have a different expected output file for ORCA and
non-ORCA. And allows for more structured errors anyway.

Use the new function for the case of trying to read from a WRITABLE
external table. There was no test for that in the main test suite
previously. There was one in the gpfdist suite, but that's not really the
right place, as that error is caught the same way regardless of the
protocol. While we're at it, re-word the error message and change the error
code to follow the Postgres error message style guide.