[#140601033] Optimize equivalence classes intersection
This commit request introduces a better algorithm for generating the intersection
of two sets of equivalence classes. The input sets are arrays of column
reference sets which denotes individual equivalence classes.

Since the equivalence classes are disjoint by definition, the implementation
could be reduced from quadratic time to the linear time by utilizing an HashMap.

We build a hash-map using the columns as keys and equivalence classes as the
value.

Below is a sample input and output scenario.

`Classes1`: `[(a,b),(c,d,e),(f,g)]`
`Classes2`: `[(a),(b,c,d),(e),(f,g)]`

The hashmap after first iteration on `Classes1`:
`a->(a,b)`
`b->(a,b)`
`c->(c,d,e)`
`d->(c,d,e)`
`e->(c,d,e)`
`f->(f,g)`
`g->(f,g)`

In the probe iteration, once we detect an intersection, we replace the intersected
columns' entry in the HashMap with an empty set to avoid duplication.

In the probe iteration, first `(a)` will be read first and column `a` will be
probed. `map[a] = (a,b)`. So the intersection of `(a,b)` and `(a)` is `(a)` and it will
be added to result list. `a`'s entry in the map is emptied. Hash map after this step
will look like below:
`a->()`
`b->(a,b)`
`c->(c,d,e)`
`d->(c,d,e)`
`e->(c,d,e)`
`f->(f,g)`
`g->(f,g)`

Then, `(b,c,d)` will be processed. First column `b` is used to probe the hash map and
it will return `(a,b)`. The intersection will be `(b)` and it will be appended to the
result list. `b`'s entry in the map is emptied. Hash Map after this step:
`a->()`
`b->()`
`c->(c,d,e)`
`d->(c,d,e)`
`e->(c,d,e)`
`f->(f,g)`
`g->(f,g)`

Then column `c` will be looked up and hash map will return `(c,d,e)`. The intersaction
of `(b,c,d)` and `(c,d,e)` will be `(c,d)` and it will be added to result list. `c` and
`d`'s entry is also emptied in the Hash Map after this step and it will be like below:
`a->()`
`b->()`
`c->()`
`d->()`
`e->(c,d,e)`
`f->(f,g)`
`g->(f,g)`

So, when we look up for column `d` at this point which is the last column of `(b,c,d)`,
since `d`'s entry is blank in the hash map, it will not produce a duplicate intersection
result of `(c,d)` again.

The final result will be `[(a),(b),(c,d),(e),(f,g)]`

Signed-off-by: Taylor Vesely <tvesely@pivotal.io>