Add more tests for CREATEEXTTABLE privileges.
These tests are from the legacy cdbfast test suite, from
writable_ext_tbl/test_* files. They're not very exciting, and we already
had tests for some variants of these commands, but looking at the way all
these privileges are handled in the code, perhaps it's indeed best to have
tests for many different combinations. The tests run in under a second, so
it's not too much of a burden.

Compared to the original tests, I removed the SELECTs and INSERTs that
tested that you can also read/write the external tables, after successfully
creating one. Those don't seem very useful, they all basically test that
the owner of an external table can read/write it. By getting rid of those
statements, these tests don't need a live gpfdist server to be running,
which makes them a lot simpler and faster to run.

Also move the existing tests on these privileges from the 'external_table'
test to the new file.